build: Move dev-cache to ../rmgs-dev-cache
Docker sets root permissions on folders it mounts from the dev cache, including the pgdata folders

and yarn-cache. The linter being used, xo, throws an EACCES error because it tries to parse this

directory, even when it has been configured to ignore it. This is not an issue during testing, since

that typically happens in a container without the dev cache folder. The problem occurs when atom or

other editors call xo dev machine itself, to provide real time linting feedback. Moving the dev

cache to the parent folder seems to be the best solution for now.